Challenge

+++ src/main/java/com/example/letsGo/dao/RecordDao.java
@@ -1,5 +1,6 @@
 package com.example.letsGo.dao;
 
+import com.example.letsGo.domain.record.Record;
 import jakarta.persistence.Entity;
 import lombok.*;
 
@@ -7,8 +8,8 @@ import java.util.List;
 
 public interface RecordDao {
     Record addRecord(Record record);
-    Record updateRecord(int recordId, Record record);
-    void deleteRecord(int recordId);
-    Record getRecord(int recordId);
+    Record updateRecord(Long recordId, Record record);
+    void deleteRecord(Long recordId);
+    Record getRecord(Long recordId);
     List<Record> getAllRecords(int userId);
 }
